Chapter 14 add Exercise 10

+++ Chapter14-Solution/Exercise08to19/Screen.cs
@@ -6,14 +6,14 @@ namespace Exercise08to19
 {
     class Screen
     {
-        private int size = 0;
+        private double size = 0;
         private string colours = null;
         public Screen()
         {
             this.size = 0;
             this.colours = null;
         }
-        public Screen(int size, string colours)
+        public Screen(double size, string colours)
         {
             this.size = size;
             this.colours = colours;
